fix: add back multiword vue rule and rename files
