+++ examples/vue/.eslintrc.cjs
@@ -13,6 +13,6 @@ module.exports = {
     'vue/setup-compiler-macros': true,
   },
   rules: {
-    'vue/multi-word-component-names': 'off',
+    'vue/multi-word-component-names': 'error',
   },
 };
