+++ examples/vue/src/locations/AppConfigLocation.vue
