+++ examples/vue/src/locations/DialogLocation.vue
