+++ examples/vue/src/locations/EntryLocation.vue
