+++ examples/vue/src/locations/FieldLocation.vue
