+++ examples/vue/src/locations/PageLocation.vue
