+++ examples/vue/src/locations/SidebarLocation.vue
