+++ examples/vue/src/main.ts
@@ -2,12 +2,12 @@ import { createApp } from 'vue';
 
 import { init, locations } from '@contentful/app-sdk';
 
-import AppConfig from './locations/AppConfig.vue';
-import Field from './locations/Field.vue';
-import Sidebar from './locations/Sidebar.vue';
-import Entry from './locations/Entry.vue';
-import Dialog from './locations/Dialog.vue';
-import Page from './locations/Page.vue';
+import AppConfigLocation from './locations/AppConfigLocation.vue';
+import FieldLocation from './locations/FieldLocation.vue';
+import SidebarLocation from './locations/SidebarLocation.vue';
+import EntryLocation from './locations/EntryLocation.vue';
+import DialogLocation from './locations/DialogLocation.vue';
+import PageLocation from './locations/PageLocation.vue';
 import LocalhostWarning from './components/LocalhostWarning.vue';
 
 if (process.env.NODE_ENV === 'development' && window.self === window.top) {
@@ -16,12 +16,12 @@ if (process.env.NODE_ENV === 'development' && window.self === window.top) {
 } else {
   init((sdk) => {
     const locationsMap = {
-      [locations.LOCATION_ENTRY_SIDEBAR]: Sidebar,
-      [locations.LOCATION_PAGE]: Page,
-      [locations.LOCATION_ENTRY_EDITOR]: Entry,
-      [locations.LOCATION_ENTRY_FIELD]: Field,
-      [locations.LOCATION_DIALOG]: Dialog,
-      [locations.LOCATION_APP_CONFIG]: AppConfig,
+      [locations.LOCATION_ENTRY_SIDEBAR]: SidebarLocation,
+      [locations.LOCATION_PAGE]: PageLocation,
+      [locations.LOCATION_ENTRY_EDITOR]: EntryLocation,
+      [locations.LOCATION_ENTRY_FIELD]: FieldLocation,
+      [locations.LOCATION_DIALOG]: DialogLocation,
+      [locations.LOCATION_APP_CONFIG]: AppConfigLocation,
     };
 
     // Select a component depending on a location in which the app is rendered.
